evolve/requirements.txt
torch>=2.0
numpy
scikit-learn
scipy
google-generativeai>=0.8.0
